requirements.txt
gym==0.21.0
x_transformers==0.27.1
tokenizers
torch
hydra-core
dm-tree
Pillow
transformers
kornia
minedojo
black
isort
pylint
tianshou=0.4.6.post1

